requirements.txt
discord.py>=2.3.0
requests>=2.31.0
python-chess>=1.10.0
pillow>=10.0.0
svglib>=1.5.0
reportlab>=4.0.0
python-dotenv>=1.0.0
